% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(return_data_paths = FALSE)
}
\arguments{
\item{return_data_paths}{Logical. If FALSE (default), the function will return the paths
to the data files. If TRUE, the function will try to parse the data.}
}
\value{
A list of the parsed data specified in "/in/input.json"
}
\description{
This package is intended to be used inside a docker container to parse
    data from "/in/input.json" with the help of
    the input configuration in "/src/tool.yml".
    The data can be parsed by file extension.
    If return_data_paths is set to FALSE, the function will return the paths
    to the data files. If set to TRUE, the function will try to parse the data.
}
\examples{
\dontrun{
     data <- get_data()
     }
}
