% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters()
}
\value{
A list of the parsed parameters specified in "/in/parameters.json"
}
\description{
This package is intended to be used inside a docker container to parse
    parameters from "/in/parameters.json" with the help of
    the parameter configuration in "/src/tool.yml".
}
\examples{
\dontrun{
     params <- get_parameters()
     }
}
